perf(stream_data_saver_and_sender): Remove logger.debug. Rename file to stream_data_saver_and_sender.py
